Fix the middleware usage and use a proper structure
move binary exec into bin folder

use express middleware as main file

move agendash api into lib folder

update README for middleware usage

update app path for binary file

update structure to add future middleware adapters

update README for other middlewares
