+++ lib/agendash.js
