+++ package.json
@@ -2,8 +2,8 @@
   "name": "agendash",
   "version": "0.0.5",
   "description": "Agenda Dashboard",
-  "main": "agendash.js",
-  "bin": "agendash-standalone.js",
+  "main": "app.js",
+  "bin": "bin/agendash-standalone.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
